[IFRT] Define Layout and common subclasses
`ifrt::Layout` is a first-class type in IFRT describing how elements of an
array shard are laid out in memory.

Similar to existing `DType`, `Shape`, and `Sharding`, it is subclassable with
RTTI and serializable.

---

The IFRT Layout design explicitly distinguishes between a (device) default
layout and a custom layout:

* A default layout represents a runtime-chosen layout that varies by the
runtime and underlying hardware configuration. Expressing default layouts is
always done through a final class `ifrt::DefaultLayout`.

* A custom layout is a user-specified layout that is unaffected by any runtime
or hardware configuration. Expressing custom layouts are always done through
subclasses of `ifrt::ConcreteLayout`.

  * Note on custom vs. concrete: All custom layouts are a concrete layout, but not
all concrete layouts are necessarily custom layouts. A default layout may be
resolved into a concrete layout, but *using* this concrete layout will make the
layout treated as a custom layout.

This design is to accommodate the fact that the user in principle does not have
sufficient information to calculate the default layout, so the user cannot
expect a default layout to match a custom layout chosen by the user. The user
is asked to explicitly use a custom layout for arrays and computations in a
consistent way if the user requires a consistent layout between them.

---

The layout would be used in this style in IFRT APIs:

When the user provides a layout to IFRT API, `CustomLayoutRef` (`absl_nullable
std::shared_ptr<ConcreteLayout>`) is expected. The user can use `nullptr` to
request the runtime to derive in the default layout.

When the runtime returns the layout of an array or executable I/O, `LayoutRef`
(`absl_nonnull std::shared_ptr<Layout>`) will be used, to be able to indicate a
specific default layout that can be compared against other default layouts and
custom layouts.

---

This change does not use `ifrt::Layout` yet in any IFRT APIs. Such changes will
be done incrementally in subsequent changes.

A few subclasses with XLA are predefined in this change:
* `ifrt::ByteStridesLayout`: Represents a layout that can use NumPy-style byte strides to express the byte offset of array shard elements.
* `ifrt::PjRtLayout`: Wraps `xla::PjRtLayout`.
* `ifrt::BasicStringArrayLayout`: Handles a special layout of `ifrt::BasicStringArray` that uses a dense pointer of `absl::Cord*`.

There is a utility function that converts `ifrt::Layout` or
`ifrt::ConcreteLayout` into `xla::PjRtLayout`, since many users of IFRT use
`xla::PjRtLayout`.

PiperOrigin-RevId: 673120132